requirements.txt
qrcode==7.3.1
Django==3.2.7
Pillow==8.4.0
python-dotenv==0.19.1
